# docs/actions.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AutoLog Technical Blog API",
    "version": "0.3.0",
    "description": "Write complete Markdown technical articles from conversations in your AI client. Search existing topics before creating or updating. No server OpenAI API is called."
  },
  "paths": {
    "/api/categories": {
      "get": {
        "tags": [
          "categories"
        ],
        "summary": "List existing categories before classifying",
        "operationId": "listIngestCategories",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CategoryRead"
                  },
                  "type": "array",
                  "title": "Response Listingestcategories"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiToken": []
          }
        ]
      }
    },
    "/api/articles": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Save a complete Markdown article written by the client",
        "operationId": "createArticle",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticleRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "Search titles, summaries and Markdown; filter category, tag, status or exact slug",
        "operationId": "listArticles",
        "security": [
          {
            "ApiToken": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "slug",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Slug"
            }
          },
          {
            "name": "category_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category Id"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "draft",
                    "published"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ArticlePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/tags": {
      "get": {
        "tags": [
          "articles"
        ],
        "summary": "List Tags",
        "operationId": "listTags",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array",
                  "title": "Response Listtags"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiToken": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "ArticleCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1,
                "pattern": "^[a-z0-9가-힣]+(?:-[a-z0-9가-힣]+)*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "category": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "summary": {
            "type": "string",
            "maxLength": 5000,
            "minLength": 1,
            "title": "Summary"
          },
          "content_markdown": {
            "type": "string",
            "maxLength": 200000,
            "minLength": 1,
            "title": "Content Markdown"
          },
          "tags": {
            "items": {
              "type": "string",
              "maxLength": 100,
              "minLength": 1
            },
            "type": "array",
            "maxItems": 30,
            "title": "Tags"
          },
          "source_type": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Source Type",
            "default": "manual"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "related_articles": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Related Articles"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "title": "Status",
            "default": "draft"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "category",
          "summary",
          "content_markdown"
        ],
        "title": "ArticleCreate"
      },
      "ArticlePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ArticleSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "ArticlePage"
      },
      "ArticleRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "title": "Status"
          },
          "content_markdown": {
            "type": "string",
            "title": "Content Markdown"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "source_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Reference"
          },
          "related_articles": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Related Articles"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "category_id",
          "category",
          "subcategory",
          "summary",
          "tags",
          "created_at",
          "updated_at",
          "status",
          "content_markdown",
          "source_type",
          "source_reference",
          "related_articles"
        ],
        "title": "ArticleRead"
      },
      "ArticleSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "category_id": {
            "type": "string",
            "format": "uuid",
            "title": "Category Id"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "subcategory": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subcategory"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "published"
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "slug",
          "category_id",
          "category",
          "subcategory",
          "summary",
          "tags",
          "created_at",
          "updated_at",
          "status"
        ],
        "title": "ArticleSummary"
      },
...
      "CategoryRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "CategoryRead"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiToken": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [
    {
      "url": "https://autolog.example.com"
    }
  ]
}